+++ movie-api/src/main/java/com/movie/movieapi/config/WebConfig.java
@@ -1,6 +1,7 @@
 package com.movie.movieapi.config;
 
 import com.movie.movieapi.config.interceptor.DataFetchRateLimitInterceptor;
+import com.movie.movieapi.config.interceptor.DataFetchRateLimiterRedisInterceptor;
 import com.movie.movieapi.config.interceptor.ReservationRateLimitInterceptor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
@@ -14,9 +15,11 @@ public class WebConfig implements WebMvcConfigurer {
     private final DataFetchRateLimitInterceptor dataFetchRateLimitInterceptor;
     private final ReservationRateLimitInterceptor reservationRateLimitInterceptor;
 
+    private final DataFetchRateLimiterRedisInterceptor dataFetchRateLimiterRedisInterceptor;
+
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(dataFetchRateLimitInterceptor)
+        registry.addInterceptor(dataFetchRateLimiterRedisInterceptor)
                 .addPathPatterns("/api/movies/**", "/api/schedules/**");
 
         registry.addInterceptor(reservationRateLimitInterceptor)
