+++ movie-domain/domain/src/main/java/com/movie/domain/common/RateLimiterHandler.java
@@ -1,0 +1,13 @@
+package com.movie.domain.common;
+
+public interface RateLimiterHandler {
+
+    boolean isRequestLimitExceeded(String ip);
+
+    boolean isIpBlocked(String ip);
+
+    void blockIp(String ip);
+
+    boolean tryAcquire(String key);
+
+}
