+++ movie-infrastructures/redis/build.gradle
@@ -9,4 +9,5 @@ dependencies {
     implementation 'org.redisson:redisson-spring-boot-starter:3.17.7'
 
     implementation project(':movie-domain:common')
+    implementation project(':movie-domain:domain')
 }
